refactor datalib and io submodules
