+++ main.py
@@ -1,0 +1,4 @@
+from radarpipeline import Project
+
+project = Project(input_data="config.yaml")
+project.fetch_data()
